+++ examples/svelte/hello-world/src/App/themes.js
@@ -1,0 +1,2 @@
+export {default as dark} from './dark.css';
+export {default as light} from './light.css';