+++ packages/core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.34](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.33...v0.0.1-alpha.34) (2019-05-17)
+
+
+### Bug Fixes
+
+* **core:** fix styled function ([543a290](https://github.com/lttb/reshadow/commit/543a290))
+
+
+
+
+
+## [0.0.1-alpha.31](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.30...v0.0.1-alpha.31) (2019-05-17)
+
+
+### Features
+
+* **babel, core:** support extracting of styles in styled ([5dfd5c4](https://github.com/lttb/reshadow/commit/5dfd5c4))
+
+
+
+
+
 ## [0.0.1-alpha.30](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.28...v0.0.1-alpha.30) (2019-05-16)
 
 **Note:** Version bump only for package @reshadow/core