+++ packages/reshadow/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.35](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.34...v0.0.1-alpha.35) (2019-05-19)
+
+**Note:** Version bump only for package reshadow
+
+
+
+
+
+## [0.0.1-alpha.34](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.33...v0.0.1-alpha.34) (2019-05-17)
+
+**Note:** Version bump only for package reshadow
+
+
+
+
+
+## [0.0.1-alpha.33](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.32...v0.0.1-alpha.33) (2019-05-17)
+
+**Note:** Version bump only for package reshadow
+
+
+
+
+
+## [0.0.1-alpha.32](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.31...v0.0.1-alpha.32) (2019-05-17)
+
+
+### Bug Fixes
+
+* **reshadow:** add @reshadow/svelte dependency ([2d4fe40](https://github.com/lttb/reshadow/commit/2d4fe40))
+
+
+
+
+
+## [0.0.1-alpha.31](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.30...v0.0.1-alpha.31) (2019-05-17)
+
+
+### Features
+
+* **svelte:** support svelte ([db6aee2](https://github.com/lttb/reshadow/commit/db6aee2))
+
+
+
+
+
 ## [0.0.1-alpha.30](https://github.com/lttb/reshadow/compare/v0.0.1-alpha.28...v0.0.1-alpha.30) (2019-05-16)
 
 **Note:** Version bump only for package reshadow